Disable indentkeys to prevent spurious reindentation
This setting has a habit of reindenting the current line (equivalent to `==`) I'm working on, usually after the indentation was already right. For example, after typing a ternary or JSX tag.

I'd love to find the correct workaround, but meantime this effectively kills the constant reindentation of the current line.